+++ components/shopify/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/shopify",
-  "version": "0.3.18",
+  "version": "0.3.19",
   "description": "Pipedream Shopify Components",
   "main": "shopify.app.mjs",
   "keywords": [
